Introduce DataLoader
If you use *quotes* query to retreive multiple quotes and its authors, you can
see several selects to the database in the log (put `debug: true` to database
config). This is because of quote type and its author resolver which will
always make a database query.

This issue (N + 1 problem) is solved with DataLoader, which batch multiple
requests into one.